Migrated GovernorControllerTests to Xunit from NUnit
